requirements.txt
spotipy
python-dotenv
requests
# For development purposes, you might want to include: